possibled fix for sqlite Operationl Error in MATCH
Worked for Oli but not for Bogdan. Apparently some difference between sqlite
and how it handles None in statements such as "x MATCH y" (where y might be
None).

Or it is a LANG problem (althugh LANG=C did not help).

Using the function form MATCH(y,x) seemed to work in a test case.