updated docs: make clear that Gibbs free energy is computed
We run all MD simulations in the NPT ensemble and hence the computed solvation
free energy is a Gibbs free energy. Pre 0.6.0 we ran the FEP in NVT and so the
solvation free energy was technically a Helmholtz free energy (although the
actual difference is small).

Note that older simulations (before this commit) stored the solvation energy
under the names Helmholtz *and* Gibbs (for historic reasons). There is no
built-in way to check if FEP simulations were run in NVT or NPT (it depends
on the template MDP file) so it is up to the user to verify if the calculated
free energy is Helmholtz or Gibbs. From *this* commit onwards, the default
process in MDPOW always uses NPT for FEP and the Helmholtz variable was
removed. However, if a user provides a custom fep.mdp file that runs NVT
simulations then the free energy will be wrongly labeled as a Gibbs energy.
It is assumed that a user who figures out how to use custom MDP files will
also be smart enought to interpret the output correctly...

Closes #22.