fixed: distance for initial box was not set by default
If the distance was not given explicitly as setup.distance, the default distance from
equil.DIST was not chosen. Note that the DIST[water] default was set to 1.0 again (instead
of 1.2) because for right now we want to be consistent with the SAMPL5 simulations.
We might set it to 1.2 later (to reduce problems with cutoffs in the box for small molecules)